sdl2 app class

+++ src/game-field.cpp
@@ -13,7 +13,7 @@ void GameField::addShip(int numberOfDecks, int sx, int sy, Direction dir)
 	int dx[] = {-1, 1, 0, 0};
 	int dy[] = {0, 0, -1, 1};
 	for (int i = 0; i < numberOfDecks; i++) {
-		field[sy + dy[dir] * i][sx + dx[dir] * i] = '#';
+		field[sy + dy[(int)dir] * i][sx + dx[(int)dir] * i] = '#';
 	}
 }
